Date: 4/08/2025 commit

+++ src/main/java/com/ecommerce/belgro/Model/Cart.java
@@ -24,7 +24,7 @@ public class Cart {
     private Set<CartItem> cartItems = new HashSet<>();
     private double totalSellingPrice;
     private int totalItem;
-    private double totalMrpPrice; //MRP stands for Manufacturing Resource Planning. It is a computerized system to manage the production and inventory of goods. MRP can be used in different types of industries like manufacturing, retail, and logistics.
+    private int totalMrpPrice; //MRP stands for Manufacturing Resource Planning. It is a computerized system to manage the production and inventory of goods. MRP can be used in different types of industries like manufacturing, retail, and logistics.
     private int discount;
     private String couponCode;
 }
